+++ CHANGELOG.md
@@ -7,6 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+* Edit Structure macro should now edit the base actor and not the unlinked token actor
+
+
 ## [2.1.2] - 2025-01-29
 
 ### Fixed